* fix fadvice and madvice inclusion ... again * fix rrdcgi warning * fix warnings in tcl bindings  -- Peter Breitenlohner peb mppmu.mpg.de
git-svn-id: svn://svn.oetiker.ch/rrdtool/branches/1.2/program@1268 a5681a0c-68f1-0310-ab6d-d61299d08faa

+++ src/rrd_cgi.c
@@ -1287,7 +1287,7 @@ s_var **rrdcgiReadVariables(void)
 	if (i<numargs) {
 
 	    /* try to find out if there's already such a variable */
-	    for (k=0; k<i && (strncmp (result[k]->name,cp, esp-cp) || !(strlen (result[k]->name) == esp-cp)); k++);
+	    for (k=0; k<i && (strncmp (result[k]->name,cp, esp-cp) || !(strlen (result[k]->name) == (size_t)(esp-cp))); k++);
 
 	    if (k == i) {	/* No such variable yet */
 		if ((result[i] = (s_var *)malloc(sizeof(s_var))) == NULL)
